requirements.txt
python-dotenv~=1.1.0
rich~=14.0.0
mysql-connector-python~=9.3.0
dependency-injector~=4.46.0